Explain how to run this
For somebody completely new to tide and sqlx, it is a bit much to go through both projects to try to understand what kind of environment variables to set in order to get this to work. I suppose most people just stop and look for something else, so I added a simple explanation on how to actually run this.